Overhaul file structure to be compatible with React Router, update regionSelector component
