+++ client/App.jsx
@@ -1,19 +1,21 @@
 import React from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 // import './stylesheets/App.scss';
-import SavedStretches from './components/savedStretches';
-import RegionSelector from './components/regionSelector';
-import StretchDisplay from './components/stretchDisplay';
-import Routines from './components/routines';
+import { PrivateRoutes } from './components/privateRoutes';
+import Home from './components/home';
+import Login from './components/login';
 
 const App = () => {
   return (
-    <>
-      <p>Hello world</p>
-      <RegionSelector />
-      <StretchDisplay />
-      <SavedStretches />
-      <Routines />    
-    </>
+    <Router>
+      <Routes>
+        <Route element={<PrivateRoutes/>}>
+          <Route path='/' element={<Home/>} />  
+        </Route>
+
+        <Route path='/login' element={<Login/>} />   
+      </Routes>
+    </Router>
   )     
 }
