+++ client/stylesheets/regionSelector.scss
@@ -1,3 +1,41 @@
-img {
-    max-width: 888px;
+.container {
+    height: 600px;
+    width: 600px;
+    border-radius: 4px;
+    border: 1px solid black;
+    background-color: #92D293;
+    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2), 0 6px 20px 0 rgba(0,0,0.19);
+}
+
+.bodyFront {
+    height: 450px;
+    width: 450px;
+}
+
+.bodyFront img {
+    width: 100%;
+    height: 100%;
+}
+
+button {
+    cursor: pointer;
+    border: none;
+}
+
+.inactiveButton {
+    background: transparent;
+}
+
+.activeButton {
+    background-image: radial-gradient(red, yellow, green);
+}
+
+#chest {
+    width: 50px;
+    height: 30px;
+    position: absolute;
+    margin-top: 115px;
+    margin-left: 200px;
+    border-radius: 15px;
+    
 }
